+++ install.sh
@@ -117,6 +117,10 @@ module.exports = {
             ,"notyf" : `./${debugPath}/plugins/notyf/notyf.es.js`
             ,"sweetalert2": `./${debugPath}/plugins/sweetalert2/src/sweetalert2.js`
             ,"@shopify/draggable": `./${debugPath}/plugins/@shopify/draggable`
+            ,"prism-es6": `./${debugPath}/plugins/prism-es6/prism.js`
+            ,"nouislider": `./${debugPath}/plugins/nouislider/dist/nouislider.mjs`
+            ,"vanilla-lazyload": `./${debugPath}/plugins/vanilla-lazyload/dist/esm/lazyload.js`
+            ,"countup.js": `./${debugPath}/plugins/countup.js/dist/countup.js`
           }
         }
       ]